+++ weblate/checks/tests/test_chars_checks.py
@@ -317,3 +317,7 @@ class PuctuationSpacingCheckTest(CheckTestCase):
         self.test_failure_1 = ("string", "string!", "")
         self.test_failure_2 = ("string", "string\u00A0? string;", "")
         self.test_failure_3 = ("string", "string\u00A0; string?", "")
+
+    def test_fr_ca(self):
+        self.do_test(True, ("string", "string!", ""), "fr")
+        self.do_test(False, ("string", "string!", ""), "fr_CA")